HG: Add NULL pointer test before member address use
The current code relies on na_init_info being the first member of the
structure to avoid testing if hg_init_info is NULL. (If
hg_init_info is NULL, &hg_init_info->na_init_info is also NULL).

This is recognized as undefined behavior for some compilers and static
analysis tool. This adds a test to avoid running into the issue with
those tools.